move object id field
