+++ tests/test_typing.py
@@ -25,7 +25,7 @@ def test_it_uses_correct_types() -> None:
         datetimefield = fields.DateTimeField()
         datefield = fields.DateField()
         complexdatetimefield = fields.ComplexDateTimeField()
-        embeddeddocumentfield = fields.EmbeddedDocumentField()
+        embeddeddocumentfield = fields.EmbeddedDocumentField(Image)
         objectidfield = fields.ObjectIdField()
         genericembeddeddocumentfield = fields.GenericEmbeddedDocumentField()
         dynamicfield = fields.DynamicField()
@@ -72,7 +72,7 @@ def test_it_uses_correct_types() -> None:
     assert_type(doc.datetimefield, Optional[datetime])
     assert_type(doc.datefield, Optional[date])
     assert_type(doc.complexdatetimefield, Optional[datetime])
-    assert_type(doc.embeddeddocumentfield, Optional[str])
+    assert_type(doc.embeddeddocumentfield, Optional[Image])
     assert_type(doc.objectidfield, Optional[str])
     assert_type(doc.genericembeddeddocumentfield, Optional[str])
     assert_type(doc.dynamicfield, Optional[str])
